REFACTORIZED product detail
